Compiled and parsed dataframes
This is analysis pipeline all the way up to compilation of raw and parsed .csv files.

The MyExpt_t#d.csv files are the output from cell profiler. The df_raw.csv is compiled and has the time parameter added. The df_parsed.csv file is after parsed using the random tree classifier.