python analysis initial
This is analysis that uses python to anaylze the output images from cell profiler. It includes a script for annotating images and generating a random tree classifier. It also has script to iterate over the original cell profiler images and create new images that inclued only the parsed objects. It also includes script for easier 'manually' segmenting sube cellular regions. Finally it includes working version of script that will iterate over images in which the sub cellular images are segmented and analyze fluorescent features in the original features